pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "excel-eval"
version = "0.1.0"
description = "Comprehensive evaluation tool for AI-generated Excel files"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "jinja2>=3.1.0",
    "Pillow>=10.0.0",
    "pyyaml>=6.0",
    "xlsxwriter>=3.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
excel-eval = "excel_eval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["excel_eval"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
